Adding Logger factory
